Fix unique error messages in Send Trans. file
https://trello.com/c/NfozGoQ6

We overlooked when we completed the [Send File To S3 service](https://github.com/DEFRA/sroc-charging-module-api/pull/293) that we were sending notifications to Errbit that contained unique error messages, for example, `Error sending file /tmp/nalai50003.dat to bucket file-uploads-pre-sroc-service-gov-uk`.

The reason we avoid using unique errors messages to so [Errbit](https://github.com/errbit/errbit) can group the same errors. It has the concept of unique errors, with multiple reported instances. If our error messages are unique it leads to multiple 'error' in Errbit all with a single reported instance.

This can make it difficult to track, triage and prioritise errors. So, this change updates what we pass to the notification to make the errors standard but still pass unique credentials for the instance.

<details>
<summary>Errbit screenshot</summary>

<img width="813" alt="Screenshot 2021-03-22 at 17 37 07" src="https://user-images.githubusercontent.com/1789650/112033436-51fd1480-8b35-11eb-8e1e-34e3fb68a0a8.png">

</details>