Fix 500 error for empty region in POST transaction
A regression has crept in most likely due to the work we did resolving the deadlock issue and switching to using `UPSERT` as part of `POST` a transaction.

If you send a `POST /v2/wrls/bill-runs/{id}/transactions` with an empty `region` it causes the service to error and a `500` to be returned.

After checking, it's because the `CreateTransactionBillRunValidationService` is relying on `region` to always be populated when in this scenario its `undefined`.

This change fixes the issue.