Add new Notifier class
The intention is we'll use this in the plugin to hold a reference to the Airbrake notify method and the Hapi-pino logger. We'll then add an instance of this to the request in the same way we do with a `BillRunModel` and `RegimeModel` to make it available to the controllers and underlying services.

This is also the first instance of us not doing something in a 'service'. We need to add an 'instance' of something to `request.app`. Our convention for services are

- they are static classes
- they expose a single `go()` method

Needing to have an instance rules using a 'service' out. Our convention for models are they represent something persisted to the DB. So, again that rules a 'model' out. Typically projects will use a `lib/` folder for business or domain logic that doesn't fit neatly elsewhere.

Our `Notifier` class fits that bill. We just need to be careful `lib/` doesn't become that draw of random items everyone has in their home!