Use alternate private variable method
You can use `#myVariable` to mark a class property as private. However, we didn't clock this is only a [stage 3 proposal](https://github.com/tc39/proposal-private-methods).

This means it gets flagged by standardjs as invalid syntax and [won't be supported](https://github.com/standard/standard/issues/1401) until it reaches Stage 4 and ESLint adds support for it.

So, for now we'll just use the existing convention for declaring private class properties.