Add new Notifier plugin (#324)
Recently, we added support for sending files to AWS S3. This task will be done in the background, whilst the API ensures a response is sent to the client.

Should the process fail, we still want to see an error logged in Errbit. So, we also ensured the service was able to do this. In review, we also agreed it would be useful to see any errors recorded in the log. For this, the service would also need access to the `logger`, for example, as used by `GenerateBillRunService`.

We then had another thought; wouldn't it be nice if we could see bill run timings recorded in Errbit!? 😁

This has all coalesced into the idea that what we really need is something that makes logging and sending notifications to Errbit accessible in one place.

So, behold the **Notifier** plugin which aims to do just that!

This first change will add the new plugin. We'll migrate across to using it in subsequent changes.