Return compensation charge as boolean in /invoices
Spotted during testing by WRLS. In the `GET /v2/wrls/bill-runs/{id}/invoices/{id}` endpoint we are returning a number of fields as booleans; `credit`, `subjectToMinimumCharge`, `minimumChargeAdjustment` etc. But `compensationCharge` we are returning as a string.

We know the reason, we have inherited the the concept of holding all regime values in generic fields in the DB as strings, for example, `regime_value_9` can be used to hold a value specific to WRLS that means something completely different to CFD. The idea is to avoid exploding the number of columns in your tables.

What we have not done is cater for this when returning a response for WRLS. Having checked V1 it returns this as a boolean so we should be doing the same, if nothing else it makes the response consistent.