Add Hapi request ID to Notifier
Each [Hapi request](https://hapi.dev/api/?v=20.1.2#request) is assigned a unique ID that is a mix of actual details about the host and environment plus unique values.

For example, when running the service locally using Docker an ID will look like `1617652737349:533c1e381364:1315:kn50o8hf:10000`.

Currently, we assume the link between our log or error messages and the logged request based on context and timing. But if the `Notifier` instance we add to the request is aware of the request ID, it can ensure this is always included so we can be certain what links with what.

This change ensures whenever Notifier is used it includes the request ID in the output.