Remove transaction level status (#340)
We added a status field to the transaction based on [V1 of the API](https://github.com/DEFRA/charging-module-api) right back at the start of V2.

However, as we have iterated we have ensured that the bill run status is the primary indicator for it and all things linked to it. We have now had confirmation from the WRLS team that the status field on the transaction _is_ ignored (currently displayed in `GET /invoices/{id}`).

So, this change is about removing the defunct field.