Simplify the VSCode tasks (#341)
The recent PR [Add new docker helpers for local development](https://github.com/DEFRA/sroc-charging-module-api/pull/314) expanded the number of tasks we have in our VSCode tasks file.

On reflection, we felt that there were now commands not being used, or we were always running the same set of tasks. So, the first part of this change covers unifying some of the tasks, especially those that allow for repeated running.

The second part is being able to clearly distinguish the tasks when displayed in the list. Shorter commands that have a more direct link to what underlying command they are running we felt would help improve the understanding and use of the tasks.