Move delete bill run to background task
https://trello.com/c/S4MvN9Ev

It was found during testing that when we have large bill runs (the largest being in the 9-10K transaction range) it takes a long time for the service to respond. Our own testing confirmed we were seeing response times in excess of 20 seconds for bill runs of 10K transactions and 2K invoices.

We have imposed a requirement that all of the API's endpoints should respond in 1 second or less (on average) so this clearly falls outside that. This change tweaks the delete bill run endpoint so that sends an immediate 204 response if the bill run can be deleted, and then does the deletion in the background.