Update delete bill run service
As part of testing this we could see that during the period a bill run is being deleted it can still be accessed and returned when the view bill run endpoint is hit.

So, to protect ourselves and the users we make it clear what is going on by setting the bill run's status to `deleting` before returning the `204` response to the client.

The delete is kicked off at the sametime and will eventually complete.