Update based on feedback
@StuAA78 had an awesome suggestion. By adding `await` to the `this._deleteBillRun(billRun)` call we could drop the need to sleep in our tests.

To still get the immediate response from the endpoint we just needed to drop the `await` from the `DeleteBillRunService.go(req.app.billRun)` in the controller.

On top of that we could re-jig where we use await in the service. Key thing to note, just calling `billRun.$query().patch({ status: 'deleting' })` doesn't actually seem to get it to invoke. Only by putting the `return` or an `await` in front does it actually seem to run!