Move test sleep() method to GeneralHelper
A small bit of housekeeping. We recently made a change where we thought we might also need to use the `sleep()` method in another test. So, we refactored it to the `GeneralHelper` where it makes better sense to be.

Turns out we didn't need it but we still feel there is value in this minor bit of housekeeping. Hence the change.