Next Customer File Reference Service (#342)
https://trello.com/c/CSrOSDTP/1942-s-develop-generate-customer-files-v2

As part of this we need to be able to generate file references for customer files. This change amends the existing "next file" functionality to allow for separate transaction file and customer file references.

It introduces a BaseNextFileReferenceService which does most of the heavy lifting, along with NextTransactionFileReferenceService and NextCustomerFileReferenceService which inherit from the base service and set the filename and db field accordingly.

Along with this, the existing file_number field in the sequence_counters table is renamed to transaction_file_number and a new customer_file_number field is introduced.