Presenters for writing customer files (#347)
https://trello.com/c/CSrOSDTP/1942-s-develop-generate-customer-files-v2

This branch creates presenters which will transform data in the customer table into the format required for customer files.

It also introduces PresenterHelper with a generateNumberedColumns method which can be used to generate the arrays of column names (eg. ['col01', 'col02', 'col03'] used in all of the file presenter tests, and refactors the existing transaction file tests to use it.