Send Customer File admin endpoint (#349)
https://trello.com/c/CSrOSDTP/1942-s-develop-generate-customer-files-v2

Weekly file generation is handled by a regular job which hits an endpoint, triggering a call to SendCustomerFileService -- this change implements the PATCH /admin/{regimeId}/customers which will do this. It uses a stub SendCustomerFileService which will be overwritten once the actual service has been reviewed and merged.

This change also introduces a new lib, StaticLookup. This is intended to hold static data which is used by different services, presenters etc. The first bit of static data to be added is the regions array. A future PR will introduce this in existing services where necessary.