Exclude zero value transactions from transaction files (#353)
https://trello.com/c/RBEmt0JV/1924-include-required-content-in-transaction-files-v2

It was found during testing that when generating transaction files, we were excluding zero-value invoices from the file but weren't explicitly excluding zero-value transactions as per AC4. This change implements this.

Note that we would normally have created a modifier on the model for this, however modifiers are only available to Objection and not Knex; this means we can't use a modifier here as the streaming interface we rely on for generating files only works with Knex queries.