Delete invoice error handling (#356)
After we moved delete invoice to a background task, we found that trying to delete a non-existent invoice, or an invoice not connected to the given bill run, would bring the service down. This turned out to be caused by FetchAndValidateBillRunInvoiceService throwing an error in these circumstances -- moving the delete invoice task to the background meant that there was nothing to catch these errors.

In order to resolve this, we have moved FetchAndValidateBillRunInvoiceService into the controller, so that a 404 or 409 error will be presented to the user if it occurs. We have also amended DeleteInvoiceService to use the notifier to log any additional errors which may occur so that they don't bring down the service.