Fix transaction file compensation charge defect (#360)
https://trello.com/c/RBEmt0JV/1924-include-required-content-in-transaction-files-v2

Certain fields in the transaction file are only populated depending on whether or not the transaction is a compensation charge. We found during testing that this was not being done correctly. This turned out to be because the compensation charge is stored in generic field regimeValue17 which is stored as a string, not a boolean, and therefore was always determined to be true regardless of whether it was the string 'true' or 'false'. We fix this by adding a helper method to correctly parse this field, and amend the existing _blankIfCompensationChargeOrMinimumCharge and _blankIfNotCompensationCharge methods to use this instead of regimeValue17.

This led to our GenerateTransactionFileService tests all failing, and the revelation that the service did not actually include regimeValue17 in its select query! This has now been resolved.