Use optional chaining to guard against null value (#362)
Our fix for the transaction file compensation charge defect was found to cause issues for bill runs with minimum charge transactions. This turned out to be because regimeValue17 is stored as null for minimum charges, thus causing an error when we attempt to make it lower case. Adding the optional chaining operator to our _compensationCharge method means we now return false without issue if regimeValue17 is null.

This defect was undiscovered in unit test as our test incorrectly stated regimeValue17 would be 'false' for a minimum charge adjustment. The test has been amended accordingly.