Create customer files table (#364)
https://trello.com/c/LwTT9wKp/1948-m-develop-generate-customer-files-audit-table-v2

This change creates the customer_files table. It differs slightly from the v1 table, in that we don't record the status -- we may adjust this in future but for now we work on the assumption that files are only added to the db after being successfully generated and sent (whereas v1 adds them with an initialised status, changing to exported once they are generated and sent).