Add generate customer file task
https://trello.com/c/Nu0BAjJE

Client systems can send us changes to customer details which we then send across to SOP. Customers changes are grouped by region and if a bill run with a matching region is `sent` we also generate that customer file.

But if there happens to be no bill runs for a region that week, the business still wants us to ensure the customer changes are sent to SOP. We do that by triggering the creation of the customer files on a schedule once a week. The plan is for a Jenkins job to be set up to run on a schedule and use this task we are adding to do that.

This is the first for the service; a 'service' that is called outside of a request (in this case from `package.json`). To be consistent with the ECS naming convention and what we have done in Jenkins for our migrate DB and seed DB actions, we are also 'namespacing' these services under 'tasks'.

So, this change adds the first such service task which when called will run the `SendCustomerFileService` and generate the files for all regions.