Create exported customers table (#369)
https://trello.com/c/LwTT9wKp/1948-m-develop-generate-customer-files-audit-table-v2

Once a customer change is exported, we want to delete the customer's details as per the acceptance criteria but retain a record that their details were updated. The exported_customers table created in this PR will hold this record by saving customer_reference alongside customer_files.id of the customer file that the details change was included in.