Fix SendCustomerFileService error logging (#373)
https://trello.com/c/Nu0BAjJE

Whilst looking to [Add generate customer file task](https://github.com/DEFRA/sroc-charging-module-api/pull/370) we came across some odd behaviour.

We are working on a new `TaskNotifier` intended for use outside of a Hapi request. To confirm it was working as expected we introduced an error into the `SendCustomerFileService` as it processed each regime. So, we should have seen 4 errors reported to Errbit but we only saw 1.

After much (_much_ 😩) digging we tracked down the problem. Airbrake runs each 'notification' through a series of filters before deciding whether to forward it on. We were falling foul of a check which looks to see if the same error has been reported in the last 1000ms. In our case it had.

```javascript
    let s = JSON.stringify(notice.errors);
    if (s === lastNoticeJSON) {
      return null;
    }
```

_https://github.com/airbrake/airbrake-js/blob/master/packages/browser/src/filter/debounce.ts for the filter and https://github.com/airbrake/airbrake-js/blob/a1663636301e256c61480306ce92f512f8075bed/packages/browser/src/base_notifier.ts#L153-L160 for where it's checked in the `notify()` call._

It ignores the additional data we add which in our case makes each error unique. So, we have fallen foul of our attempts to ensure errors are _not_ distinct to allow for easier grouping in Errbit 🤦.

In most places that convention is still fine. It's just in this case there is a possibility that errors are generated for each region, or across the 4 regimes in a process that takes less than 1000 in total.

So, this change tweaks our messages to make them distinct for each iteration.