Add TaskNotifier and refactor existing Notifier
https://trello.com/c/Nu0BAjJE

As aprt of [Add generate customer file task](https://github.com/DEFRA/sroc-charging-module-api/pull/370) we want to ensure we can log what is going on when the task is running plus notify Errbit if any errors happen. Also, the `SendCustomerFileService` service expects a 'notifier' to be passed in.

We couldn't use the existing `Notifier` because it depends on the **Pino** and **Airbrake** instance being passed in from the [Hapi request object](https://hapi.dev/api/?v=20.1.2#request). So, we would need to create a new one that would instantiate the **Pino** and **Airbrake** instances needed.

When we got into it, we realised a few things

- there was a lot of duplication in the 2 notifiers
- using the generic name `Notifier` was going to cause confusion
- there was no need to depend on the [Hapi request object](https://hapi.dev/api/?v=20.1.2#request) to provide our **Pino** and **Airbrake** instances

So, this change covers both add a new `TaskNotifier` and major refactoring of the Notifier classes and usage.