Remove complexity just to stub Pino
In order to test the Airbrake side of the notifications we had to resort to stubbing the BaseNotification class's `_setNotifier()` method. Having done that, it didn't make much sense to be going through various hoops and breaking our existing conventions just to support stubbing Pino.

So, this change removes all that complexity and we just stick to stubbing the BaseNotification `_setLogger()` instead.