Refactor to use existing Pino and Airbrake
Started seeing `MaxListenersExceededWarning` in the test output again

```
(node:1152) MaxListenersExceededWarning: Possible EventEmitter memory leak detected. 11 beforeExit listeners added to [process]. Use emitter.setMaxListeners() to increase limit
```

This is a result of how Airbrake adds listeners to `process` events and our moving to each instantiation of a `Notifier` adding more listeners leading to the warning being reported.

It's not likely to be an issue when the app actually runs. But as we already have these instances in the context of a request it made sense to refactor things to use them, and therefore stop the warning when the tests are run.

It did mean a slight tweak to things; what the airbrake plugin adds to Hapi is the `notify() method itself, not an instance of Airbrake's `Notifier`.