Use Airbrake Notifier instance rather than notify
Previously we were expecting an instance of the Airbrake.Notifier.notify() function to be passed into our base notifier, or we would instantiate one via `_setNotifier()`.

Lurking to the side though has been the `flush()` method. This is important for the `TaskNotifier`. In the case of a task it's a one time call; the task completes and the app is expected to close. Because of the way Airbrake works though, any error notifications are sent synchronously in the background. It is unlikely they will complete sending before the app exists which means they never make it to Errbit.

Airbrake provides the `flush()` method to deal with this. It forces all queued notifications to be sent, and if called with `await` will block the process until they are all sent. Our issue was it's a method on the Airbrake `Notifier` instance. Though we generate one either directly in the `BaseNotifier` or in the Airbrake plugin, we were actually passing around an instance of its `notify()` method.

So, to support flush and as it turns out be consistent with the Pino logger (we pass around a Pino `logger` instance not it's `info()` function) this change refactors everything to expect a `Notifier` instance.

Note - this does break other stuff which will be fixed in the next commit!