Update Airbrake plugin and how it sets things up
If you review how the plugin was implemented prior to this change most of the functionality was actually focused on supporting adding the Airbrake `notify()` method as a Hapi [server method](https://hapi.dev/api/?v=20.1.2#-servermethods).

Because it was setup twice, we had moved the promise resolution functions outside of the instantiation so they could be reused. With the need to store the instance of Airbrake `Notifier` in the server, not the `notify()` method it gave us the opportunity to review and refactor the plugin.

Now we know all logging and notifications will be done using the `RequestNotifier` instance we add to the Hapi [request object](https://hapi.dev/api/?v=20.1.2#request) we have re-written it with that intent in mind.