Fix airbrake health controller
With the change to the Airbrake plugin and how it sets things up we needed to update the airbrake health controller which is intended to test Airbrake is setup correctly.

We've stuck with using Airbrake via the `server.app.notification` instance instead of using the `RequestNotifier`, as this is intended to check Airbrake and we would prefer to go direct to test the connection rather than deal with issues possibly added by the `RequestNotifier` code.