Use pino for airbrake failures
Now we are more familiar with how things work plus Pino and Airbrake we had the realisation that _should_ Airbrake fail to send a notification we could stick to using Pino for the log entry. This would help keep our logs consistent as JSON output and more easily searchable.

So, whilst we have been making a major change to the plugin now seems a good time to also do this.