Add TaskRunner
This combined with the `TaskRunner.go(process.argv[2])` call in `app/tasks/index.js` is our solution for centralising all the code for running tasks and not duplicating that code within the tasks themselves.

So, the process will be

- add your new task as a standard JS class using the 'service' pattern we have adopted
- add a reference to it in `app/tasks/index.js`, which is what we do for all our other classes
- then add a reference to it in `TaskRunner` as well as it's switch statement
- finally add your script to `package.json` being sure to pass the class name as an argument

Ideally we wouldn't have to keep opening up `TaskRunner` each time we add a task. But we don't intend to do it that often. `TaskRunner` though manages instantiating the notifier (which it will do properly once `TaskNotifier` is merged), flushing it for Airbrake errors, and then exiting when the work is done.