Add generate customer file task (#370)
https://trello.com/c/Nu0BAjJE

Client systems can send us changes to customer details which we then send across to SOP. Customers changes are grouped by region and if a bill run with a matching region is `sent` we also generate that customer file.

But if there happens to be no bill runs for a region that week, the business still wants us to ensure the customer changes are sent to SOP. We do that by triggering the creation of the customer files on a schedule once a week. The plan is for a Jenkins job to be set up to run on a schedule and use this task we are adding to do that.

This is the first for the service; a 'service' that is called outside of a request (in this case from `package.json`). To be consistent with the ECS naming convention and what we have done in Jenkins for our migrate DB and seed DB actions, we are also 'namespacing' these services under 'tasks'.

So, this change adds the first such service task which when called will run the `SendCustomerFileService` and generate the files for all regions.

** Notes on TaskRunner

It combined with the `TaskRunner.go(process.argv[2])` call in `app/tasks/index.js` is our solution for centralising all the code for running tasks and not duplicating that code within the tasks themselves.

So, the process will be

- add your new task as a standard JS class using the 'service' pattern we have adopted
- add a reference to it in `app/tasks/index.js`, which is what we do for all our other classes
- then add a reference to it in `TaskRunner`
- finally, add your script to `package.json` being sure to pass the class name as an argument

Ideally, we wouldn't have to keep opening up `TaskRunner` every time we add a task. But we don't intend to do it that often. `TaskRunner` though manages instantiating the notifier (which it will do properly once `TaskNotifier` is merged), flushing it for Airbrake errors, and then exiting when the work is done.