Add TaskNotifier and refactor existing Notifier (#379)
https://trello.com/c/Nu0BAjJE

As part of [Add generate customer file task](https://github.com/DEFRA/sroc-charging-module-api/pull/370) we want to ensure we can log what is going on when the task is running plus notify Errbit if any errors happen. Also, the `SendCustomerFileService` service expects a 'notifier' to be passed in.

We couldn't use the existing `Notifier` because it depends on the **Pino** and **Airbrake** instance being passed in from the [Hapi request object](https://hapi.dev/api/?v=20.1.2#request). So, we would need to create a new one that would instantiate the **Pino** and **Airbrake** instances needed.

When we got into it, we realised a few things

- there was a lot of duplication in the 2 notifiers
- using the generic name `Notifier` was going to cause confusion
- our use of the Notifier classes made redundant a lot of what the Airbrake plugin is doing

So, this change covers both adding a new `TaskNotifier` and major refactoring of the Notifier classes and usage to make logging and sending errors to Errbit consistent in the project.

* Notes on changes to the Airbrake plugin

If you review how the plugin was implemented prior to this change most of the functionality was actually focused on supporting adding the Airbrake `notify()` method as a Hapi [server method](https://hapi.dev/api/?v=20.1.2#-servermethods).

Because it was set up twice, we had moved the promise resolution functions outside of the instantiation so they could be reused. With the need to store the instance of Airbrake `Notifier` in the server to support our refactoring instead of the `notify()` method, it gave us the opportunity to review and refactor the plugin.

Now with all logging and notifications being done using the `RequestNotifier` instance, we add to the Hapi [request object](https://hapi.dev/api/?v=20.1.2#request) we have re-written it with that intent in mind.