Update TaskRunner to use TaskNotifier
When we added the [customer file task](https://github.com/DEFRA/sroc-charging-module-api/pull/370) we included a temporary implementation for a 'notifer' because we knew [Add TaskNotifier and refactor existing Notifier](https://github.com/DEFRA/sroc-charging-module-api/pull/379) was waiting in the wings.

The `TaskNotifier` has now been merged so this change updates the `TaskRunner` to use it.