Send Customer File Service moves customers to exported_customers (#384)
https://trello.com/c/LwTT9wKp/1948-m-develop-generate-customer-files-audit-table-v2

This change adds MoveCustomersToExportedTableService to SendCustomerFileService so that customer records are moved from customers to exported_customers when a customer file is generated.

We also make a small change to one of the tests for MoveCustomersToExportedTableService. This test was sometimes failing as it relied on the tested items being in a specific order in an array; the change means the test is now independent of the array order.