Fix model table names
When starting to write tests for a new `ShowCustomerFileService` we started getting errors that their was something wrong with the customer files table name.

We've been using it for awhile so this didn't make sense. So, we started comparing the `CustomerFileModel` and `ExportedCustomerModel` to other models. Then it hit; in these models we had declared the tableName using snake_case.

Looking at `BillRunModel` and `AuthorisedSystemModel`, both of which have multi-word names, their table names are declared as camel case.

When we updated the table names in `CustomerFileModel` and `ExportedCustomerModel` to be camelCase the error went away. So this change corrects any table name declarations which have been declared using snake_case.

** Note - The error only appears when you try and access a child relationship as we were doing. As we have yet no reason to query for the sequence counters linked to a regime we have not had an error with `SequenceCounterModel`, which is how that model has been getting away with it.