Add show customer file endpoint
We needed to make some changes to the test file after this change. As we now are following the convention of stubbing any underlying services (they have been tested elsewhere) we initially hit an error that the final test always failed because the stub was always returning a result.

We then spotted this is because the tests had been setup on the basis that we stub `JsonWebToken` just once and then restore just once at the end.

So, we needed to tweak everything so that we stub beforeEach() and restore afterEach().