Fix previous changes in customer file export (#386)
During testing we found that if a customer file was successfully generated but sending failed (eg. an invalid bucket is specified), the customer records in that file would be incorrectly included in the next customer file. This change prevents this happening:

- Once a customer file record is created, we patch the customer changes for that regime and region so that customer_file_id equals the id of the created record, ensuring we do not overwrite the file id if it already exists for a customer change;
- When we generate the file, we only select customer changes that have the correct customer_file_id.

This is implemented by refactoring GenerateCustomerFileService so that we pass the instance of CustomerFileModel to it, which it uses to get the id (plus other things needed for the file, such as region). We also refactor SendCustomerFileService to match.