Change unit-test to test
We never ran the old `test` script. Our current setup with docker negates the need for the migratedb (plus we prefer to not hide it in there). We also run linting and tests separately in our CI.

So, this change ditches the old test command and uses what we were doing in `unit-test`. Now if someone runs `npm test` it will do what they expect. Plus we can start using `npm test` ourselves.