Use consistent naming convention package scripts (#388)
We are currently going through a refactor of our Jenkins jobs and one of the things noted was the inconsistency in the naming of the tasks in our `package.json`.

We started with all lowercase and no spaces (based on what we had in V1). But we also have some kebab-case.

Before we cement in our Jenkins jobs as part of the production environment setup, we want to take this opportunity to apply a consistent naming convention for them.

Based on this [stackoverflow answer](https://stackoverflow.com/a/45536339/6117745) which also aligns with our preferences, we'll update everything to use kebab-case.

We also want to do a bit of a tidy up which includes using more standard names and removing duplication.