Move plugin based services to own folder
> Part of a series of housekeeping tasks to tidy up the project

We know the number of files in the `app/services` folder is getting a little out of hand so we have decided to start grouping them either by 'resource', for example, `BillRun` or 'function', for example, `plugins`.

This change moves any service whose primary purpose is to support a plugin into a new `plugins` folder.