Add db reset package.json script
We have the ability to roll back migrations through the existing scripts. We can then follow that up with a reseed of the DB.

Locally that is relatively simple to do through chaining the commands we pass through (see the VSCode task `🗄️ DB (CMA)` for an example).

But we could also do with this ability in our AWS environments (in non-production) to support testing. Ideally to make setting up the Jenkins job easier this would be kicked off from a single `npm run` script.

So, this change implements that run script and updates the existing VSCode tasks to use it.