Fix jwk file selection for INTEGRATION (#391)
We recently updated our Jenkins jobs to remove the enormous amount of duplication in them. One of the key config items is the environment config.

It's used by the app to determine which key file to verify incoming requests against, but also in the config of the ECS task definitions.

When the `ENVIRONMENT` value was hardcoded into the task definition it was fine to have one value for ECS, and another for our app. But now both need to reference the same value, the confusion about `(INT)EGRATION` vs `(TRA)INING` has raised its head.

So this change updates the solution to use the naming convention currently in place in our AWS environments which is `TRA`.