Move JwtStrategyAuth to app/lib folder (#390)
> Part of a series of housekeeping tasks to tidy up the project

Don't ask us why but right back at the start of the project when we were implementing the authentication logic we assumed there would be more 'auth' related code.

Turns out we just need this one little module, so putting it in a folder all on its lonesome is a little overkill. This change moves it to the `app/lib` folder so it finally has some friends! 🤗