Add missing 'await' to test (#392)
We found that the tests for MoveCustomersToExportedTableService would very occasionally fail. On closer examination, it looked like there was a missing await in the test setup which was probably causing a situation where customer details weren't in place ready for test, which would subsequently fail. This change hopes to fix this!